requirements.txt
Django>=4.2
djangorestframework
reportlab
Django>=4.2
djangorestframework
reportlab
